Independent change of step and zstep in controlframe
Hello,

first of all thank you for your great work.

I want to propose some changes in the behaviour of the controlframe for
jogging in x/y and z direction IN THE CASE that zstep is enabled in the
.bcnc file.

1. Independent INC/DEC/MUL/DIV Buttons for x/y and z
In the moment the values step and zstep are only individually changeable
by the combobox. I think it might be better if they are also
individually changeable by the buttons.

2. shift of the z-related buttons to the right side of the frame
So I can insert the new buttons on the most right side without changing
the rest of the frame too much. There are some "|" symbols to show when
x/y and z are seperated.

3. New button for move to a most up z-position
I use this often to clear y workspace on the CNC maschine. I use "G53
Z-2" which will work fine after "homing". For this button I also
inserted a new UNICODE symbol to unicode.py.