Z enter Rapid -  Driller
Added Z order rapid enter, descend at fast speed until surface = surface + 1.00
Added to cut, drill command.
In the driller plugin modified to start in the second movement (the first movement corresponds to the completion of the previous block) that generates drilling twice in the same place.
I do not have a laser, check operation.
In my version of pc, I also generate holes; If these modifications deserve to be merged, and they think it is useful, I will add it in the next intervention